ziva the diva

+++ Engine/Ball.h
@@ -10,8 +10,8 @@ public:
 	void Draw(Graphics& gfx);
 	Color c{ 255,255,0 };
 
-	float x = Graphics::ScreenWidth / 2;
-	float y = Graphics::ScreenHeight / 2;
+	float x = float(Graphics::ScreenWidth / 2);
+	float y = float(Graphics::ScreenHeight / 2);
 	static constexpr int w = 20;
 	static constexpr int h = 20;
 	float vx = 1.0f;
